First changes for adopting Grocy 3.1.0
This changes are the biggest ones yet in the history of this app, since the changed data types (Grocy 3.1.0 doesn't send strings for everything anymore) are in nearly every aspect of the app.

This is the first batch of changes, but it only works slightly.
It will need much more testing and some tricks (to handle the possible nil, empty string or number returns) to make it run as before. But the code should be much cleaner now.