Starting of doing selected action after Quick add
This seems quite buggy, since SwiftUI produces non-logical behaviour. Sometimes it works, sometimes it doesn't.
It's not great, but this is all effort I will put into this, since I already spent too much time debugging this ;-)
Since this is optional, there shouldn't be a big problem with this.
It should be the final fix for implementing #117