Added rudimentary reminders integration
Since other popular shopping list apps sync their shopping list to reminders, why shouldn't Grocy Mobile do it?

All it needs is an example from the Apple guide and ...
Well, it's not ready yet. But it can add the shopping list items to a list in reminders. But it needs some work, so the functionality is hidden.