Change the way picture are presented
Originally, the App got the link for a picture and pushed it to the AsyncImage component from SwiftUI. The problem with this was, that there was no authentication for this component, so tis always failed for non-demo servers.

This now changes the way the pictures get presented, by fetching data via API and then showing it in an Image. This should fix the long-standing issue #113.